+++ rmq_metadata.py
@@ -55,6 +55,8 @@
             # Directory name for log files.
             log_dir = "DIRECTORY_PATH/logs"
             # File name to program log.
+            # Note:  Name will be changed to include the exchange name being
+            #   processed.
             log_file = "rmq_metadata.log"
             # Directory name for archived messages.
             # Must be set if archive in any of the queues is set to True.
@@ -229,6 +231,7 @@ def validate_create_settings(cfg, **kwargs):
     status_flag = True
     base_dir = gen_libs.get_base_dir(__file__)
 
+    # Check on non-processed messages directory.
     if not os.path.isabs(cfg.message_dir):
         cfg.message_dir = os.path.join(base_dir, cfg.message_dir)
 
@@ -239,6 +242,7 @@ def validate_create_settings(cfg, **kwargs):
         err_msg = err_msg + msg
         status_flag = False
 
+    # Check on log files directory.
     if not os.path.isabs(cfg.log_dir):
         cfg.log_dir = os.path.join(base_dir, cfg.log_dir)
 
@@ -254,6 +258,7 @@ def validate_create_settings(cfg, **kwargs):
         err_msg = err_msg + msg
         status_flag = False
 
+    # Check on archived messages directory.
     if cfg.archive_dir and not os.path.isabs(cfg.archive_dir):
         cfg.archive_dir = os.path.join(base_dir, cfg.archive_dir)
 
@@ -265,6 +270,7 @@ def validate_create_settings(cfg, **kwargs):
             err_msg = err_msg + msg
             status_flag = False
 
+    # Check on temporary message processing directory.
     if not os.path.isabs(cfg.tmp_dir):
         cfg.message_dir = os.path.join(base_dir, cfg.tmp_dir)
 
@@ -275,6 +281,10 @@ def validate_create_settings(cfg, **kwargs):
         err_msg = err_msg + msg
         status_flag = False
 
+    # Check on file entries.
+    status_flag, err_msg = _validate_files(cfg, status_flag, err_msg)
+
+    # Check on final directory for each queue.
     for queue in cfg.queue_list:
         status, msg = gen_libs.chk_crt_dir(queue["directory"], write=True,
                                            read=True, no_print=True)
@@ -284,6 +294,53 @@ def validate_create_settings(cfg, **kwargs):
             status_flag = False
 
     return cfg, status_flag, err_msg
+
+
+def _validate_files(cfg, status_flag, err_msg, **kwargs):
+
+    """Function:  _validate_files
+
+    Description:  Private function for validate_create_settings.  Validates the
+        file entries in the configuration file.
+
+    Arguments:
+        (input) cfg -> Configuration settings module for the program.
+        (input) status_flag -> True|False - successfully validation.
+        (input) err_msg -> Error message from checks.
+        (output) status_flag -> True|False - successfully validation.
+        (output) err_msg -> Error message from checks.
+
+    """
+
+    # Check on Stanford NLP language module file.
+    if not os.path.isabs(cfg.lang_module):
+        msg = "lang_module not set to absolute path: %s" % (cfg.lang_module)
+        err_msg = err_msg + msg
+        status_flag = False
+
+    else:
+        status, msg = gen_libs.chk_crt_file(cfg.lang_module, read=True,
+                                            no_print=True)
+
+        if not status:
+            err_msg = err_msg + msg
+            status_flag = False
+
+    # Check on Stanford NLP jar file.
+    if not os.path.isabs(cfg.stanford_jar):
+        msg = "stanford_jar not set to absolute path: %s" % (cfg.stanford_jar)
+        err_msg = err_msg + msg
+        status_flag = False
+
+    else:
+        status, msg = gen_libs.chk_crt_file(cfg.stanford_jar, read=True,
+                                            no_print=True)
+
+        if not status:
+            err_msg = err_msg + msg
+            status_flag = False
+
+    return status_flag, err_msg
 
 
 def non_proc_msg(rmq, log, cfg, data, subj, r_key, **kwargs):